refactor: simplify how Commander works and add a new option in Controller
Commanders now just receive an array and do all the job of running the three main functions. The Commander class per scene now it's clear and can be used as a god class if necessary, for some game general stuff.

The new function on Controller helps with Composition, can be used to receive a single ScriptableObject with all data. So Player has a SO, that can be sent to Movement, Combat, Input, etc, by their OnAwakes (and still a single SO entry point, for data, pretty cool).